extracted muiStyled wip
