+++ packages/material-ui/src/styles/index.d.ts
@@ -23,6 +23,7 @@ export {
   StyledComponentProps,
 } from './withStyles';
 export { default as withTheme, WithTheme } from './withTheme';
+export { default as muiStyled } from './muiStyled';
 export { default as styled, ComponentCreator, StyledProps } from './styled';
 export {
   createGenerateClassName,
