+++ packages/material-ui/src/styles/index.js
@@ -10,6 +10,7 @@ export * from './transitions';
 export { default as useTheme } from './useTheme';
 export { default as withStyles } from './withStyles';
 export { default as withTheme } from './withTheme';
+export { default as muiStyled } from './muiStyled';
 export {
   createGenerateClassName,
   jssPreset,
