+++ packages/material-ui/src/styles/muiStyled.d.ts
@@ -1,0 +1,8 @@
+/**
+ * Cutom styled functionality that support mui specific config.
+ *
+ * @param options Takes an incomplete theme object and adds the missing parts.
+ * @returns A complete, ready to use theme object.
+ */
+// TODO: fix typings
+export default function adaptV4Theme(component: any, config: any, muiConfig: any): React.Component;
